CHANGE users dedicated urls so that they all reside in a subdirectory (_).
robots.txt disallow browsing of /_/ to crawlers.

Not that web crawler would explore actual users information, but in
the current situation, we create a new user for every crawler that browses
the site, every time it browses it.

This is non sense so to have a better indexation, we should avoid
crawlers to do so.

Just in case it is not clear enough: this is not a security fix, it is a bug fix.